Massive commit - see details below
I'm in the process of recreating/adapting worlds and creating
special menus for various parts of the site (docs, tutorial, etc.).
The work is not completed but there are so many changes to be
done that it makes sense to not wait until the end in case
some file corruption occurs.